Fixed casting warnings from strtoul on macOS, change port to uint16_t.

+++ util.h
@@ -48,19 +48,19 @@ char *get_url_scheme(const char *url, const char *default_scheme);
 char *get_url_from_host(const char *scheme, const char *host);
 
 // Get port from host
-int32_t get_host_port(const char *host, size_t host_len, int32_t default_port);
+uint16_t get_host_port(const char *host, size_t host_len, uint16_t default_port);
 
 // Strip and parse port from host
-int32_t strip_host_port(char *host, size_t host_len, int32_t default_port);
+uint16_t strip_host_port(char *host, size_t host_len, uint16_t default_port);
 
 // Strip ipv6 brackets from host
 bool strip_host_ipv6_brackets(char *host);
 
 // Use scheme based on port specified
-const char *get_port_scheme(int32_t port, const char *default_scheme);
+const char *get_port_scheme(uint16_t port, const char *default_scheme);
 
 // Get default port for a scheme
-int32_t get_scheme_default_port(const char *scheme);
+uint16_t get_scheme_default_port(const char *scheme);
 
 // Convert proxy list returned by FindProxyForURL to a list of uris separated by commas.
 char *convert_proxy_list_to_uri_list(const char *proxy_list, const char *default_scheme);
